Renaming package names to azure
