+++ extensions/http-client-vertx/runtime/src/main/resources/META-INF/services/com.azure.core.http.vertx.VertxProvider
@@ -1,1 +1,1 @@
-io.quarkiverse.azureservices.core.http.vertx.runtime.QuarkusVertxProvider
+io.quarkiverse.azure.core.http.vertx.runtime.QuarkusVertxProvider
