--- extensions/storage-blob/deployment/src/main/java/io/quarkiverse/azureservices/storage/blob/deployment/StorageBlobProcessor.java
@@ -1,55 +1,0 @@
-package io.quarkiverse.azureservices.storage.blob.deployment;
-
-import org.jboss.jandex.ClassInfo;
-import org.jboss.jandex.DotName;
-
-import io.quarkiverse.azureservices.storage.blob.runtime.StorageBlobServiceClientProducer;
-import io.quarkus.arc.deployment.AdditionalBeanBuildItem;
-import io.quarkus.deployment.annotations.BuildProducer;
-import io.quarkus.deployment.annotations.BuildStep;
-import io.quarkus.deployment.builditem.CombinedIndexBuildItem;
-import io.quarkus.deployment.builditem.ExtensionSslNativeSupportBuildItem;
-import io.quarkus.deployment.builditem.FeatureBuildItem;
-import io.quarkus.deployment.builditem.IndexDependencyBuildItem;
-import io.quarkus.deployment.builditem.nativeimage.ReflectiveClassBuildItem;
-
-class StorageBlobProcessor {
-
-    static final String FEATURE = "azure-storage-blob";
-
-    @BuildStep
-    FeatureBuildItem feature() {
-        return new FeatureBuildItem(FEATURE);
-    }
-
-    @BuildStep
-    AdditionalBeanBuildItem producer() {
-        return new AdditionalBeanBuildItem(StorageBlobServiceClientProducer.class);
-    }
-
-    @BuildStep
-    ExtensionSslNativeSupportBuildItem activateSslNativeSupport() {
-        return new ExtensionSslNativeSupportBuildItem(FEATURE);
-    }
-
-    @BuildStep
-    IndexDependencyBuildItem indexDependency() {
-        return new IndexDependencyBuildItem("com.azure", "azure-storage-blob");
-    }
-
-    @BuildStep
-    void reflectiveClasses(CombinedIndexBuildItem combinedIndex, BuildProducer<ReflectiveClassBuildItem> reflectiveClasses) {
-        String[] modelClasses = combinedIndex
-                .getIndex()
-                .getKnownClasses()
-                .stream()
-                .map(ClassInfo::name)
-                .map(DotName::toString)
-                .filter(n -> n.startsWith("com.azure.storage.blob.implementation.models.")
-                        || n.startsWith("com.azure.storage.blob.models."))
-                .sorted()
-                .toArray(String[]::new);
-        reflectiveClasses.produce(new ReflectiveClassBuildItem(false, true, modelClasses));
-    }
-
-}
