+++ integration-tests/src/test/java/io/quarkiverse/azure/storage/blob/it/StorageBlobResourceIT.java
@@ -1,0 +1,7 @@
+package io.quarkiverse.azure.storage.blob.it;
+
+import io.quarkus.test.junit.QuarkusIntegrationTest;
+
+@QuarkusIntegrationTest
+public class StorageBlobResourceIT extends StorageBlobResourceTest {
+}
