--- integration-tests/src/test/java/io/quarkiverse/azureservices/storage/blob/it/StorageBlobResourceIT.java
@@ -1,7 +1,0 @@
-package io.quarkiverse.azureservices.storage.blob.it;
-
-import io.quarkus.test.junit.QuarkusIntegrationTest;
-
-@QuarkusIntegrationTest
-public class StorageBlobResourceIT extends StorageBlobResourceTest {
-}
